+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FWFGeneratedWebKitApis.h
@@ -744,6 +744,12 @@ NSObject<FlutterMessageCodec> *FWFWKWebViewHostApiGetCodec(void);
 - (void)setAllowsBackForwardForWebViewWithIdentifier:(NSNumber *)identifier
                                            isAllowed:(NSNumber *)allow
                                                error:(FlutterError *_Nullable *_Nonnull)error;
+- (void)setAllowsLinkPreviewForWebViewWithIdentifier:(NSNumber *)identifier
+                                           isAllowed:(NSNumber *)allow
+                                               error:(FlutterError *_Nullable *_Nonnull)error;
+- (void)setUserAgentForWebViewWithIdentifier:(NSNumber *)identifier
+                                   userAgent:(nullable NSString *)userAgent
+                                       error:(FlutterError *_Nullable *_Nonnull)error;
 - (void)setCustomUserAgentForWebViewWithIdentifier:(NSNumber *)identifier
                                          userAgent:(nullable NSString *)userAgent
                                              error:(FlutterError *_Nullable *_Nonnull)error;