+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FWFWebViewHostApi.m
@@ -276,6 +276,14 @@
 }
 
 - (void)
+    setAllowsLinkPreviewForWebViewWithIdentifier:(nonnull NSNumber *)identifier
+                                       isAllowed:(nonnull NSNumber *)allow
+                                           error:(FlutterError *_Nullable __autoreleasing *_Nonnull)
+                                                     error {
+  [[self webViewForIdentifier:identifier] setAllowsLinkPreview:allow.boolValue];
+}
+
+- (void)
     setNavigationDelegateForWebViewWithIdentifier:(nonnull NSNumber *)identifier
                                delegateIdentifier:(nullable NSNumber *)navigationDelegateIdentifier
                                             error: