+++ packages/webview_flutter/webview_flutter_wkwebview/test/src/web_kit/web_kit_test.dart
@@ -833,6 +833,14 @@ void main() {
         ));
       });
 
+      test('setAllowsLinkPreview', () {
+        webView.setAllowsLinkPreview(false);
+        verify(mockPlatformHostApi.setAllowsLinkPreview(
+          webViewInstanceId,
+          false,
+        ));
+      });
+
       test('setCustomUserAgent', () {
         webView.setCustomUserAgent('hello');
         verify(mockPlatformHostApi.setCustomUserAgent(